requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.2
django-allauth==0.48.0
django-cors-headers==3.11.0
django-on-heroku==1.1.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.0.0
environs==9.5.0
gunicorn==20.1.0
idna==3.3
marshmallow==3.15.0
oauthlib==3.2.0
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
